Jordan branch (#5)
* Added stuff

Added gallery, settings, and home page. Also added a navigation bar at the bottom to switch between the pages

* Changes to Main and Settings

Minor detail changes to main such as positioning for the navigation bar and icons. Updated settings to have multiple features such as username and password fields that are editable and settings to make profile private and such.

---------

Co-authored-by: Keith Crawford <121456679+crawfordk99@users.noreply.github.com>